# perf(runtime): short-circuit `queue_async_op` for Poll::Ready