--- a/cli/tests/testdata/044_bad_resource.ts.out
+++ b/cli/tests/testdata/044_bad_resource.ts.out
@@ -1,2 +1,2 @@
-[WILDCARD]error: Uncaught (in promise) BadResource: Bad resource ID
+[WILDCARD]error: Uncaught [WILDCARD] BadResource: Bad resource ID
 [WILDCARD]