@aapoalas 2022-09-05
LGTM. I wonder what the negative effect on non-short async ops is.